requirements.txt
altair==4.0.0
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asn1crypto==1.5.1
asttokens==2.4.1
attrs==23.1.0
beautifulsoup4==4.9.3
blinker==1.7.0
cachetools==5.3.2
certifi==2020.12.5
cffi==1.16.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.7
comm==0.2.0
contourpy==1.2.0
cryptography==36.0.2
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
distro==1.8.0
entrypoints==0.4
exceptiongroup==1.1.3
executing==2.0.1
filelock==3.13.1
fonttools==4.44.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
greenlet==1.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.17.3
idna==2.10
importlib-metadata==6.8.0
importlib-resources==6.1.1
insta-scrape==2.1.2
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.26.1
oauthlib==3.2.2
openai==1.3.0
oscrypto==1.3.0
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
platformdirs==3.11.0
playwright==1.9.1
praw==7.2.0
prawcore==2.0.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pycparser==2.21
pycryptodomex==3.19.0
pydantic==2.5.1
pydantic_core==2.14.3
pydeck==0.8.1b0
pyee==8.1.0
Pygments==2.16.1
PyJWT==2.8.0
Pympler==1.0.1
pyOpenSSL==22.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.12.0
safetensors==0.4.0
selenium==3.141.0
selenium-stealth==1.0.6
semver==3.0.2
six==1.15.0
smmap==5.0.1
sniffio==1.3.0
snowflake-connector-python==3.4.1
snowflake-sqlalchemy==1.5.1
sortedcontainers==2.4.0
soupsieve==2.2
SQLAlchemy==1.4.50
stack-data==0.6.3
streamlit==1.28.2
tenacity==8.2.3
textblob==0.17.1
TikTokApi==3.9.1
tokenizers==0.14.1
toml==0.10.2
tomlkit==0.12.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.35.0
tweepy==4.14.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
update-checker==0.18.0
urllib3==1.26.3
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.9
websocket-client==0.57.0
wordcloud==1.9.2
zipp==3.17.0
